梳理项目结构，清理不需要的缓存文件并添加.gitignore
