+++ .gitignore
@@ -1,0 +1,7 @@
+__pycache__/generator.cpython-312.pyc
+comments/__pycache__/generator.cpython-312.pyc
+__pycache__/memeapp.cpython-312.pyc
+comments/__pycache__/memeapp.cpython-312.pyc
+./__pycache__
+*.py[cod]
+*$py.class
