--- comments/main.py
@@ -1,8 +1,0 @@
-from memeapp import MemeApp
-
-# 主程序入口判断
-if __name__ == "__main__":
-    # 实例化应用
-    app = MemeApp()
-    # 运行应用
-    app.run()
